move dropdown menu to widgets. consolidate icons.
