+++ web/components/chat/chat-message.tsx
@@ -11,7 +11,7 @@ import { MultipleOrSingleAvatars } from 'web/components/multiple-or-single-avata
 import { Modal, MODAL_CLASS } from 'web/components/layout/modal'
 import { UserAvatarAndBadge } from 'web/components/widgets/user-link'
 import Link from 'next/link'
-import DropdownMenu from 'web/components/comments/dropdown-menu'
+import DropdownMenu from 'web/components/widgets/dropdown-menu'
 import { DotsHorizontalIcon, ReplyIcon } from '@heroicons/react/solid'
 import { UserHovercard } from '../user/user-hovercard'
 import { DisplayUser } from 'common/api/user-types'
@@ -77,7 +77,9 @@ export const ChatMessageItem = memo(function ChatMessageItem(props: {
                     onClick: () => onReplyClick(chat),
                   },
                 ]}
-                icon={<DotsHorizontalIcon className="text-ink-400 h-4 w-4" />}
+                buttonContent={
+                  <DotsHorizontalIcon className="text-ink-400 h-4 w-4" />
+                }
               />
             )}
           </Row>
