+++ web/components/contract/header-actions.tsx
@@ -22,7 +22,7 @@ import { db } from 'web/lib/supabase/db'
 import { duplicateContractHref } from '../buttons/duplicate-contract-button'
 import { followMarket, unfollowMarket } from '../buttons/follow-market-button'
 import { ReportModal } from '../buttons/report-button'
-import DropdownMenu from '../comments/dropdown-menu'
+import DropdownMenu from '../widgets/dropdown-menu'
 import { Row } from '../layout/row'
 import { SweepsToggle } from '../sweeps/sweeps-toggle'
 import { getLinkTarget } from '../widgets/linkify'
